biblioteca_digital/microservicio_usuarios/requirements.txt
Django==5.0.6
djangorestframework==3.16.1
